update search param path
