+++ domain/vod/src/test/java/usecase/details/movie/GetMoviesByCategoryTest.kt
@@ -1,8 +1,8 @@
-package usecases.details.movie
+package usecase.details.movie
 
 import com.google.common.truth.Truth.assertThat
 import details.repository.MovieRepository
-import details.usecase.movie.GetSimilarMoviesByMovieId
+import details.usecase.movie.GetMoviesByCategory
 import entity.Genre
 import entity.Movie
 import exceptions.RetrievingDataFailureException
@@ -11,59 +11,57 @@ import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
 import kotlinx.datetime.LocalDate
 import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
+import kotlin.test.Test
 
-class GetSimilarMoviesByMovieIdTest{
+class GetMoviesByCategoryTest {
     private lateinit var movieRepository: MovieRepository
-    private lateinit var getSimilarMoviesByMovieId: GetSimilarMoviesByMovieId
+    private lateinit var getMoviesByCategory: GetMoviesByCategory
 
     @BeforeEach
     fun setUp() {
         movieRepository = mockk()
-        getSimilarMoviesByMovieId = GetSimilarMoviesByMovieId(movieRepository)
+        getMoviesByCategory = GetMoviesByCategory(movieRepository)
     }
-
     @Test
-    fun `execute() should return similar movies when available`() = runTest {
+    fun `execute() should return movies when available`() = runTest {
         // Given
-        val movieId = 1
-        val similarMovies = listOf(testMovie1, testMovie2)
-        coEvery { movieRepository.getSimilarMoviesByMovieId(movieId) } returns similarMovies
+        val category = Genre.ACTION
+        val movies = listOf(testMovie1, testMovie2)
+        coEvery { movieRepository.getMoviesByCategory(category) } returns movies
 
         // When
-        val result = getSimilarMoviesByMovieId.execute(movieId)
+        val result = getMoviesByCategory.execute(category)
 
         // Then
-        assertThat(result).isEqualTo(similarMovies)
+        assertThat(result).isEqualTo(movies)
     }
-
     @Test
-    fun `execute() should return empty list when there are no similar movies`() = runTest {
+    fun `execute() should return empty list when there are no movies in the category`() = runTest {
         // Given
-        val movieId = 2
-        coEvery { movieRepository.getSimilarMoviesByMovieId(movieId) } returns emptyList()
+        val category = Genre.DRAMA
+        coEvery { movieRepository.getMoviesByCategory(category) } returns emptyList()
 
         // When
-        val result = getSimilarMoviesByMovieId.execute(movieId)
+        val result = getMoviesByCategory.execute(category)
 
         // Then
         assertThat(result).isEmpty()
     }
+
     @Test
     fun `execute() should throw exception when repository fails`() = runTest {
         // Given
-        val movieId = 1
-        coEvery { movieRepository.getSimilarMoviesByMovieId(movieId) } throws RetrievingDataFailureException("Failed to retrieve similar movies")
+        val category = Genre.COMEDY
+        coEvery {
+            movieRepository.getMoviesByCategory(category)
+        } throws RetrievingDataFailureException("Failed to retrieve movies by category")
 
         // When / Then
         assertThrows<RetrievingDataFailureException> {
-            getSimilarMoviesByMovieId.execute(movieId)
+            getMoviesByCategory.execute(category)
         }
     }
-
-
-
     companion object {
         private val testMovie1 = Movie(
             id = 101,
@@ -87,5 +85,4 @@ class GetSimilarMoviesByMovieIdTest{
             overview = "Overview of movie 2"
         )
     }
-
 }
