+++ domain/vod/src/test/java/usecase/details/tv_series/GetEpisodeDetailsUseCaseTest.kt
@@ -1,7 +1,8 @@
-package usecase
+package usecase.details.tv_series
 
 import details.repository.TvSeriesRepository
 import details.usecase.tv_series.GetEpisodeDetailsUseCase
+import entity.Episode
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
@@ -24,7 +25,7 @@ class GetEpisodeDetailsUseCaseTest {
     @Test
     fun `execute() should return episode details from TvSeriesRepository`() = runTest {
         // Given
-        val expected = mockk<entity.Episode>()
+        val expected = mockk<Episode>()
         coEvery { repository.getEpisodeDetails(1, 1, 1) } returns expected
 
         // When
